+++ src/aggregation/bucket/filter.rs
@@ -7,7 +7,7 @@ use crate::aggregation::intermediate_agg_result::{
     IntermediateAggregationResult, IntermediateAggregationResults, IntermediateBucketResult,
 };
 use crate::aggregation::segment_agg_result::{CollectorClone, SegmentAggregationCollector};
-use crate::query::{AllQuery, BooleanQuery, Query, Weight};
+use crate::query::{Query, Weight};
 use crate::schema::Schema;
 use crate::{DocId, SegmentReader, TantivyError, TERMINATED};
 
@@ -46,7 +46,7 @@ impl FilterAggregation {
 
     /// Parse the stored query JSON into a Tantivy Query object
     pub fn parse_query(&self, schema: &Schema) -> crate::Result<Box<dyn Query>> {
-        parse_elasticsearch_query(&self.query, schema)
+        crate::query::parse_query(&self.query, schema)
     }
 
     /// Get the fast field names used by this aggregation (none for filter aggregation)
@@ -313,259 +313,11 @@ pub struct IntermediateFilterBucketResult {
     pub sub_aggregations: IntermediateAggregationResults,
 }
 
-/// Parse Elasticsearch query JSON into Tantivy Query objects using Tantivy's QueryParser
-pub fn parse_elasticsearch_query(
-    query_json: &serde_json::Value,
-    schema: &Schema,
-) -> crate::Result<Box<dyn Query>> {
-    use serde_json::Value;
-
-    match query_json {
-        Value::Object(obj) => {
-            if obj.is_empty() {
-                return Err(TantivyError::InvalidArgument(
-                    "Query object cannot be empty".to_string(),
-                ));
-            }
-
-            if obj.len() > 1 {
-                return Err(TantivyError::InvalidArgument(
-                    "Query object should contain exactly one query type".to_string(),
-                ));
-            }
-
-            let (query_type, query_params) = obj.iter().next().unwrap();
-            match query_type.as_str() {
-                "term" => parse_term_query_with_tantivy_parser(query_params, schema),
-                "range" => parse_range_query_with_tantivy_parser(query_params, schema),
-                "bool" => parse_bool_query(query_params, schema),
-                "match" => parse_match_query_with_tantivy_parser(query_params, schema),
-                "match_all" => Ok(Box::new(AllQuery)),
-                _ => Err(TantivyError::InvalidArgument(format!(
-                    "Unsupported query type: {}",
-                    query_type
-                ))),
-            }
-        }
-        _ => Err(TantivyError::InvalidArgument(
-            "Query must be a JSON object".to_string(),
-        )),
-    }
-}
-
-/// Parse Elasticsearch term query using Tantivy's QueryParser: { "term": { "field": "value" } }
-fn parse_term_query_with_tantivy_parser(
-    params: &serde_json::Value,
-    schema: &Schema,
-) -> crate::Result<Box<dyn Query>> {
-    use crate::query::QueryParser;
-    use crate::tokenizer::TokenizerManager;
-    use serde_json::Value;
-
-    let obj = params.as_object().ok_or_else(|| {
-        TantivyError::InvalidArgument("Term query parameters must be an object".to_string())
-    })?;
-
-    if obj.len() != 1 {
-        return Err(TantivyError::InvalidArgument(
-            "Term query must specify exactly one field".to_string(),
-        ));
-    }
-
-    let (field_name, field_value) = obj.iter().next().unwrap();
-    let field = schema.get_field(field_name).map_err(|_| {
-        TantivyError::InvalidArgument(format!("Field '{}' not found in schema", field_name))
-    })?;
-
-    // Convert the JSON value to a string for parsing
-    let value_str = match field_value {
-        Value::String(s) => {
-            if s.is_empty() {
-                // Handle empty string case - use quotes to make it a valid query
-                "\"\"".to_string()
-            } else if s.trim().is_empty() {
-                // Handle whitespace-only strings - use quotes to preserve them
-                format!("\"{}\"", s)
-            } else if s.contains(':') || s.contains('-') || s.contains('T') {
-                // Handle date/time strings and other special formats - use quotes
-                format!("\"{}\"", s)
-            } else {
-                s.clone()
-            }
-        }
-        Value::Number(n) => n.to_string(),
-        Value::Bool(b) => b.to_string(),
-        _ => {
-            return Err(TantivyError::InvalidArgument(
-                "Term query value must be string, number, or boolean".to_string(),
-            ))
-        }
-    };
-
-    // Use Tantivy's QueryParser to properly handle type conversion
-    let tokenizer_manager = TokenizerManager::default();
-    let query_parser = QueryParser::new(schema.clone(), vec![field], tokenizer_manager);
-
-    // Create a query string in the format "field:value"
-    let query_str = format!("{}:{}", field_name, value_str);
-
-    query_parser
-        .parse_query(&query_str)
-        .map_err(|e| TantivyError::InvalidArgument(format!("Failed to parse term query: {}", e)))
-}
-
-/// Parse Elasticsearch range query using Tantivy's QueryParser: { "range": { "field": { "gte": 10, "lt": 20 } } }
-fn parse_range_query_with_tantivy_parser(
-    params: &serde_json::Value,
-    schema: &Schema,
-) -> crate::Result<Box<dyn Query>> {
-    use crate::query::QueryParser;
-    use crate::tokenizer::TokenizerManager;
-    use serde_json::Value;
-
-    let obj = params.as_object().ok_or_else(|| {
-        TantivyError::InvalidArgument("Range query parameters must be an object".to_string())
-    })?;
-
-    if obj.len() != 1 {
-        return Err(TantivyError::InvalidArgument(
-            "Range query must specify exactly one field".to_string(),
-        ));
-    }
-
-    let (field_name, range_params) = obj.iter().next().unwrap();
-    let field = schema.get_field(field_name).map_err(|_| {
-        TantivyError::InvalidArgument(format!("Field '{}' not found in schema", field_name))
-    })?;
-
-    let range_obj = range_params.as_object().ok_or_else(|| {
-        TantivyError::InvalidArgument("Range query field parameters must be an object".to_string())
-    })?;
-
-    // Build a range query string in Tantivy's format
-    let mut lower_bound = None;
-    let mut upper_bound = None;
-    let mut lower_inclusive = true;
-    let mut upper_inclusive = true;
-
-    for (bound_type, bound_value) in range_obj {
-        let value_str = match bound_value {
-            Value::String(s) => s.clone(),
-            Value::Number(n) => n.to_string(),
-            Value::Bool(b) => b.to_string(),
-            _ => {
-                return Err(TantivyError::InvalidArgument(
-                    "Range query value must be string, number, or boolean".to_string(),
-                ))
-            }
-        };
-
-        match bound_type.as_str() {
-            "gte" => {
-                lower_bound = Some(value_str);
-                lower_inclusive = true;
-            }
-            "gt" => {
-                lower_bound = Some(value_str);
-                lower_inclusive = false;
-            }
-            "lte" => {
-                upper_bound = Some(value_str);
-                upper_inclusive = true;
-            }
-            "lt" => {
-                upper_bound = Some(value_str);
-                upper_inclusive = false;
-            }
-            _ => {
-                return Err(TantivyError::InvalidArgument(format!(
-                    "Unsupported range bound: {}",
-                    bound_type
-                )))
-            }
-        }
-    }
-
-    if lower_bound.is_none() && upper_bound.is_none() {
-        return Err(TantivyError::InvalidArgument(
-            "Range query must specify at least one bound".to_string(),
-        ));
-    }
-
-    // Use Tantivy's QueryParser to properly handle type conversion
-    let tokenizer_manager = TokenizerManager::default();
-    let query_parser = QueryParser::new(schema.clone(), vec![field], tokenizer_manager);
-
-    // Create a range query string in Tantivy's format: "field:[lower TO upper]"
-    let lower_str = lower_bound.as_deref().unwrap_or("*");
-    let upper_str = upper_bound.as_deref().unwrap_or("*");
-    let left_bracket = if lower_inclusive { "[" } else { "{" };
-    let right_bracket = if upper_inclusive { "]" } else { "}" };
-
-    let query_str = format!(
-        "{}:{}{} TO {}{}",
-        field_name, left_bracket, lower_str, upper_str, right_bracket
-    );
-
-    query_parser
-        .parse_query(&query_str)
-        .map_err(|e| TantivyError::InvalidArgument(format!("Failed to parse range query: {}", e)))
-}
-
-/// Parse Elasticsearch bool query: { "bool": { "must": [...], "should": [...], "must_not": [...] } }
-fn parse_bool_query(params: &serde_json::Value, schema: &Schema) -> crate::Result<Box<dyn Query>> {
-    use crate::query::Occur;
-    use serde_json::Value;
-
-    let obj = params.as_object().ok_or_else(|| {
-        TantivyError::InvalidArgument("Bool query parameters must be an object".to_string())
-    })?;
-
-    let mut subqueries = Vec::new();
-
-    for (clause_type, clause_queries) in obj {
-        let occur = match clause_type.as_str() {
-            "must" => Occur::Must,
-            "should" => Occur::Should,
-            "must_not" => Occur::MustNot,
-            _ => {
-                return Err(TantivyError::InvalidArgument(format!(
-                    "Unsupported bool query clause: {}",
-                    clause_type
-                )))
-            }
-        };
-
-        let queries_array = clause_queries.as_array().ok_or_else(|| {
-            TantivyError::InvalidArgument(format!(
-                "Bool query '{}' clause must be an array",
-                clause_type
-            ))
-        })?;
-
-        for query_json in queries_array {
-            let query = parse_elasticsearch_query(query_json, schema)?;
-            subqueries.push((occur, query));
-        }
-    }
-
-    Ok(Box::new(BooleanQuery::new(subqueries)))
-}
-
-/// Parse Elasticsearch match query using Tantivy's QueryParser: { "match": { "field": "value" } }
-/// Note: This implementation uses Tantivy's QueryParser which handles text analysis and tokenization
-fn parse_match_query_with_tantivy_parser(
-    params: &serde_json::Value,
-    schema: &Schema,
-) -> crate::Result<Box<dyn Query>> {
-    // For match queries, we can use the same logic as term queries since QueryParser handles tokenization
-    parse_term_query_with_tantivy_parser(params, schema)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::schema::{Schema, FAST, TEXT};
+    use crate::query::parse_query;
+    use crate::schema::{Schema, FAST, INDEXED, TEXT};
     use serde_json::json;
 
     #[test]
@@ -590,11 +342,11 @@ mod tests {
 
         // Test string term query
         let query_json = json!({ "category": "electronics" });
-        let _query = parse_term_query_with_tantivy_parser(&query_json, &schema).unwrap();
+        let _query = parse_query(&json!({ "term": query_json }), &schema).unwrap();
 
         // Test numeric term query
         let query_json = json!({ "price": 100 });
-        let _query = parse_term_query_with_tantivy_parser(&query_json, &schema).unwrap();
+        let _query = parse_query(&json!({ "term": query_json }), &schema).unwrap();
     }
 
     #[test]
@@ -605,7 +357,7 @@ mod tests {
         let schema = schema_builder.build();
 
         let query_json = json!({ "price": { "gte": 10, "lt": 100 } });
-        let _query = parse_range_query_with_tantivy_parser(&query_json, &schema).unwrap();
+        let _query = parse_query(&json!({ "range": query_json }), &schema).unwrap();
     }
 
     #[test]
@@ -616,15 +368,15 @@ mod tests {
 
         // Test term query
         let query_json = json!({ "term": { "category": "electronics" } });
-        let _query = parse_elasticsearch_query(&query_json, &schema).unwrap();
+        let _query = parse_query(&query_json, &schema).unwrap();
 
         // Test match_all query
         let query_json = json!({ "match_all": {} });
-        let _query = parse_elasticsearch_query(&query_json, &schema).unwrap();
+        let _query = parse_query(&query_json, &schema).unwrap();
 
         // Test invalid query type
         let query_json = json!({ "invalid_query": {} });
-        let result = parse_elasticsearch_query(&query_json, &schema);
+        let result = parse_query(&query_json, &schema);
         assert!(result.is_err());
     }
 
@@ -719,7 +471,7 @@ mod tests {
             }
         });
 
-        let _query = parse_elasticsearch_query(&query_json, &schema).unwrap();
+        let _query = parse_query(&query_json, &schema).unwrap();
         // Should parse successfully without errors
     }
 }
